cfp: purpleidea aka James
Here are three proposals in response to the CFP. I don't expect all
three to be accepted, but I wanted to present the available options. If
you'd like me to present purpleidea-2.md, then early notification of
acceptance would be helpful since this material has never been seen
before and more time to prepare would be helpful!

They're each submitted as a separate PR as requested. This is #0.

Thank you for your consideration.